pyproject.toml
[tool.poetry]
name = "cuomputer"
version = "0.1.0"
description = "The bot for Rivers Cuomo's discord server."
authors = ["Rivers Cuomo <riverscuomo@users.noreply.github.com>"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "3.7.4.post0"
async-timeout = "3.0.1"
attrs = "21.2.0"
better-profanity = "0.6.1"
black = "21.9b0"
blinker = "1.4"
blis = "0.7.9"
cachecontrol = "0.12.10"
cached-property = "1.5.2"
cachetools = "4.2.2"
catalogue = "1.0.0"
certifi = "2021.5.30"
cffi = "1.15.1"
chardet = "3.0.4"
charset-normalizer = "2.1.1"
click = "8.1.3"
colorama = "0.4.4"
commonmark = "0.9.1"
coolname = "1.1.0"
cymem = "2.0.5"
demoji = "0.2.1"
deprecated = "1.2.13"
discord2 = "2.0.1"
et-xmlfile = "1.1.0"
firebase-admin = "5.1.0"
flask = "1.1.2"
flask-caching = "1.9.0"
flask-cors = "3.0.10"
flask-login = "0.4.1"
flask-moment = "0.10.0"
flask-wtf = "0.14.3"
google-api-core = "1.34.0"
google-api-python-client = "2.29.0"
google-auth = "1.35.0"
google-auth-httplib2 = "0.1.0"
google-auth-oauthlib = "0.4.6"
google-cloud-core = "1.7.3"
google-cloud-dialogflow = "2.9.0"
google-cloud-firestore = "2.3.4"
google-cloud-storage = "1.42.3"
google-cloud-translate = "3.9.0"
google-crc32c = "1.2.0"
google-resumable-media = "2.0.3"
googleapis-common-protos = "1.57.0"
grpcio = "1.51.1"
grpcio-status = "1.48.2"
gspread = "5.7.2"
gspreader = "0.1.20"
gtts = "2.3.0"
gunicorn = "20.0.4"
h11 = "0.9.0"
h2 = "3.2.0"
hpack = "3.0.0"
hstspreload = "2021.10.1"
httpcore = "0.9.1"
httplib2 = "0.19.1"
httpx = "0.13.3"
hyperframe = "5.2.0"
idna = "2.10"
itsdangerous = "2.0.1"
jinja2 = "3.0.1"
joblib = "1.0.1"
langcodes = "3.3.0"
markupsafe = "2.0.1"
more-itertools = "8.10.0"
msgpack = "1.0.2"
multidict = "5.1.0"
murmurhash = "1.0.5"
mypy-extensions = "0.4.3"
nltk = "3.6.3"
nose = "1.3.7"
numpy = "1.23.5"
oauth2client = "4.1.3"
oauthlib = "3.1.0"
openai = "0.19.0"
openpyxl = "3.0.10"
ordered-set = "3.1.1"
ordered-set-stubs = "0.1.3"
packaging = "21.0"
pafy = "0.5.5"
pandas = "1.3.5"
pathspec = "0.9.0"
pathy = "0.10.0"
plac = "1.1.3"
platformdirs = "2.3.0"
poetry-version = "0.1.5"
polling = "0.3.2"
preshed = "3.0.5"
profanity = "1.1"
profanity-filter = "1.1.1"
proto-plus = "1.22.1"
protobuf = "3.20.3"
py-cord = "2.3.2"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pycparser = "2.21"
pydantic = "1.8.2"
pydub = "0.25.1"
pygments = "2.10.0"
pynacl = "1.5.0"
pyparsing = "2.4.7"
python-dateutil = "2.8.2"
python-dotenv = "0.21.0"
pytz = "2018.9"
random-word = "1.0.4"
redis = "3.5.3"
regex = "2021.8.28"
replicate = "0.4.0"
requests = "2.28.1"
requests-oauthlib = "1.3.0"
rfc3986 = "1.5.0"
rich = "12.6.0"
rivertils = "0.1.7"
rope = "0.20.1"
rsa = "4.7.2"
ruamel-yaml = "0.17.21"
ruamel-yaml-clib = "0.2.7"
six = "1.16.0"
smart-open = "5.2.1"
sniffio = "1.2.0"
spacy = "2.3.8"
spacy-legacy = "3.0.10"
spacy-loggers = "1.0.3"
textblob = "0.17.1"
thinc = "7.4.5"
tomli = "1.2.1"
tomlkit = "0.5.11"
tqdm = "4.62.3"
typer = "0.7.0"
types-pytz = "2022.6.0.1"
typing-extensions = "4.4.0"
uberduck = "0.0.3"
uritemplate = "4.1.1"
urllib3 = "1.25.9"
wasabi = "0.8.2"
werkzeug = "2.0.1"
wrapt = "1.13.3"
wtforms = "3.0.1"
yarl = "1.6.3"
youtube-dl = "2021.12.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
